% Generated by roxygen2 (4.0.1): do not edit by hand
\name{AdjustCovtests}
\alias{AdjustCovtests}
\title{Clean up and bonferroni adjust multiple \code{covTest} results}
\usage{
AdjustCovtests(cvtests, nvar)
}
\arguments{
\item{cvtests}{A list of results from \code{covTest}. More than one since
we fit a different model for each response variable.}

\item{nvar}{The original number of predictor variables - this info isn't
maintained in the result.}
}
\value{
An n-by-m matrix of p-values for m predictor and n response variables.
}
\description{
Clean up and bonferroni adjust multiple \code{covTest} results
}

