\docType{methods}
\name{boxplots,SPData,numeric,numeric-method}
\alias{boxplots,SPData,numeric,numeric-method}
\title{Boxplots the distribution for each channel}
\usage{
\S4method{boxplots}{SPData,numeric,numeric}(object, nrow, ncol)
}
\arguments{
  \item{nrow}{Number of rows of the boxplots to plot}

  \item{ncol}{Number of columns of boxplots to plot}
}
\description{
Boxplots the distribution for each channel
}

