% Generated by roxygen2 (4.0.1): do not edit by hand
\name{boxplots}
\alias{boxplots}
\title{Boxplots the distribution for each channel}
\usage{
boxplots(sp, nrow, ncol)
}
\arguments{
\item{nrow}{Number of rows of the boxplots to plot}

\item{ncol}{Number of columns of boxplots to plot}
}
\description{
Boxplots the distribution for each channel
}

