\docType{methods}
\name{cellClass,SPData-method}
\alias{cellClass,SPData-method}
\title{If each cell has a class (e.g. tumour or stromal) then it can be
assigned a numeric class which is retrieved through cellClass}
\usage{
\S4method{cellClass}{SPData}(object)
}
\description{
If each cell has a class (e.g. tumour or stromal) then it
can be assigned a numeric class which is retrieved through
cellClass
}

