% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{cells}
\alias{cells}
\alias{cells,SPData-method}
\alias{cells,SPData-methods}
\alias{cells<-}
\alias{cells<-,SPData-methods}
\title{Cell measurement data}
\usage{
cells(object)

cells(object) <- value

\S4method{cells}{SPData}(object)
}
\arguments{
\item{object}{The SPData object to use.}

\item{value}{The new cell measurements.}
}
\description{
Returns the n by p matrix of normalised cell measurements for
n cells and p channels.

Set the cell measurements for a given sample.
}

