\docType{methods}
\name{channelPlot,SPData,numeric-method}
\alias{channelPlot,SPData,numeric-method}
\title{Boxplots the distributions of channels depending on their class}
\usage{
\S4method{channelPlot}{SPData,numeric}(object, channel.ids)
}
\arguments{
  \item{channel.ids}{A numeric vector of channel indicies}
}
\description{
Boxplots the distributions of channels depending on their
class
}

