% Generated by roxygen2 (4.0.1): do not edit by hand
\name{clusterClass}
\alias{clusterClass}
\title{Clusters cells into different classes using
Expectation Maximisation (EM) clustering}
\usage{
clusterClass(Y, doPCA = TRUE, nclass = 2)
}
\arguments{
\item{Y}{A cell by channel matrix of readouts}

\item{doPCA}{If true, first performs PCA and uses the top
3 principle components for clustering}
}
\description{
Clusters cells into different classes using
Expectation Maximisation (EM) clustering
}

