\docType{methods}
\name{dim,SPData-method}
\alias{dim,SPData-method}
\title{Gives dimension of underlying matrix representation}
\usage{
\S4method{dim}{SPData}(x)
}
\arguments{
  \item{x}{an \R object, for example a matrix, array or
  data frame.}
}
\description{
Returns number of cells and number of proteins as dimension
of underlying matrix
}

