% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{[}
\alias{[}
\alias{[,SPData-methods}
\alias{[,SPExp-methods}
\title{Subset an SPData set}
\usage{
\S4method{[}{SPData,ANY,ANY,ANY}(x, i, j)

\S4method{[}{SPExp,ANY,ANY,ANY}(x, i)
}
\arguments{
\item{i}{Cells to subset}

\item{j}{Channels to subset}

\item{x}{The \code{SPExp} instance to subset}

\item{i}{The samples to retain}
}
\value{
An SPData object reduced to cells \code{i} and channels \code{j}

A subsetted \code{SPExp} object
}
\description{
Select SPData[i,j] for cells \code{i} and channels \code{j}.
Note that this does not subset out nearest neighbours also.

Returns a new \code{SPExp} object subsetted using the first \code{i} instances
}

