\docType{methods}
\name{findBoundary,SPData-method}
\alias{findBoundary,SPData-method}
\title{Find the cells that lie on the boundary between two classes
(currently only implemented for 2 classes)}
\usage{
\S4method{findBoundary}{SPData}(object)
}
\description{
Find the cells that lie on the boundary between two classes
(currently only implemented for 2 classes)
}

