\name{findDisjointCellSet}
\alias{findDisjointCellSet}
\title{This takes the entire set of cells and returns sets whose
members aren't neighbours of each other}
\usage{
findDisjointCellSet(sp)
}
\description{
This takes the entire set of cells and returns sets whose
members aren't neighbours of each other
}

