% Generated by roxygen2 (4.0.1): do not edit by hand
\name{FindID}
\alias{FindID}
\alias{FindIDs}
\title{Find the cell class with the highest/lowest average channel}
\usage{
FindID(sp, channel, direction = c("higher", "lower"), method = c("mean",
  "median"))

FindIDs(SPE, channel, direction = c("higher", "lower"), method = c("mean",
  "median"))
}
\arguments{
\item{sp}{An \code{SPData} object}

\item{channel}{A (partial) channel name}

\item{direction}{Whether to look for the highest or lowest average}

\item{method}{The method for average (mean or median)}

\item{SPE}{An \code{SPExp} object}
}
\value{
The class index with the higher average in the channel

A vector of cell classes corresponding to each sample.
}
\description{
Given tissue samples with different classes, it may be advantageous to single out a
class that corresponds to a certain phenotype. These are easiest found by considering
the cell class with the highest or lowest average (mean/median) expression of a given tissue,
e.g. in epithelial cells keratin is over-expressed while in stromal cells vimentin
is under-expressed.

Apply \code{\link{FindID}} across an entire \code{SPExp}
}
\details{
Most normalization routines normalize each class separately. However, this will make each
channel \eqn{N(0,1)} within a cell class, so comparison is just noise. This method normalizes
the selected channel w.r.t. cell size and concentration, then picks out the highest or
lowest average (mean/median, set by \code{method}).

\code{channel} greps the channel name to find an index containing the string, but make sure that
string exists in a channel and only once, otherwise an error will be thrown.

\code{direction} If \code{"higher"} returns the class with the highest mean/median of the selected
channel, while \code{"lower"} will return the class with the lowest mean/median.
}
\examples{
\dontrun{
## Keratin is higher in epithelial cells:
keratin.class <- FindID(sp, "Keratin", direction="higher", method="median")
}
}

