% Generated by roxygen2 (4.0.1): do not edit by hand
\name{FindOverlap}
\alias{FindOverlap}
\title{Given a list of two pathway results, find the overlap between them}
\usage{
FindOverlap(results, remove = c("same", "different"))
}
\arguments{
\item{results}{A list containg two 'pathway' matrices}

\item{remove}{Which results to keep: if \code{"same"} then 1 1 would be reported but 1 2 discarded, but if
\code{"different"} then 1 2 would be reported but 1 1 discarded (useful for finding pathways that don't
go from same component to same component)}
}
\value{
A matrix showing common pathways.
}
\description{
The result of the spatial pathway identification is a two-column matrix, where each row represents the numeric
identifiers of the causal direction of the pathway, so
1 4
3 2
would indicate there is a pathway from component 1 to 4, and from 3 to 2. If we use multiple methods to find these,
it is useful to find the overlap - pathways reported from both methods. This method does that, effectively taking the
intersection of the two matrices and returning the results in a similar matrix.
}

