\name{findTumourID}
\alias{findTumourID}
\title{Given a 2 class sample, findTumourClass returns one of the indices
(1 or 2) depending on which has the higher overall keratin concentration}
\usage{
findTumourID(sp)
}
\arguments{
  \item{sp}{An SPData object}
}
\description{
Given a 2 class sample, findTumourClass returns one of the
indices (1 or 2) depending on which has the higher overall
keratin concentration
}

