% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GeneralLassoSig}
\alias{GeneralLassoSig}
\title{Call \code{LassoSig} for multiple response variables}
\usage{
GeneralLassoSig(Y, X, B = 100, s = c("lambda.min", "lambda.1se", "halfway",
  "usefixed"), gamma.min = 0.05, fixedP = NULL, include = NULL,
  nfolds = 10, intercept = TRUE)
}
\arguments{
\item{Y}{Response matrix for general regression with \code{GeneralLassoSig}}

\item{X}{Design matrix}

\item{B}{Number of times to partition the sample}

\item{s}{The value of lambda to use in lasso. Can be:
\itemize{
\item{An actual value of lambda you have determined}
\item{\code{lambda.min} The lambda that minimises the mean squared error from n-fold cross-validation (recommended)}
\item{\code{lambda.1se} The lambda that accounts for the smallest number of parameters and is within 1 standard error from \code{lambda.min}}
\item{\code{halfway} The lambda that is halfway between \code{lambda.min} and \code{lambda.1se}}
\item{\code{usefixed} Use the smallest lambda that accounts for a given number of parameters (set by \code{fixedP})}
}}

\item{include}{Set of predictors to be force-included in OLS analysis}

\item{gamma.min}{Lower bound for gamma in the adaptive search for the best p-value (default 0.05)}

\item{fixedP}{The fixed number of parameters to use (if \code{s == "usefixed"})}

\item{nfolds}{Number of folds of cross-validation in the glmnet n-fold crossvalidation}

\item{intercept}{Whether to include an intercept in the OLS regression (default = \code{TRUE})}
}
\value{
An n-by-m matrix of p-values for n predictors and m response variables
}
\description{
Call \code{LassoSig} for multiple response variables
}
\seealso{
LassoSig
}

