% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getDir}
\alias{getDir}
\alias{getDir,SPExp-method}
\alias{getDir,SPExp-methods}
\title{Experiment directory}
\usage{
getDir(object)

\S4method{getDir}{SPExp}(object)
}
\arguments{
\item{object}{The instance of SPExp to use}
}
\description{
Returns the directory where the experiment files are located.
}

