% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{length}
\alias{length}
\title{Number of samples}
\usage{
\S4method{length}{SPExp}(x)
}
\arguments{
\item{x}{\code{SPExp instance to use}}
}
\value{
Number of samples in the \code{SPExp} instance
}
\description{
Number of samples
}

