% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lmNormalise}
\alias{lmNormalise}
\title{Provides column by column normalisation on Y
given cell size s}
\usage{
lmNormalise(Y, s, cell.classes, showPlot = FALSE)
}
\arguments{
\item{Y}{Measurement by channel matrix}

\item{s}{Vector of cell sizes}
}
\description{
Provides column by column normalisation on Y
given cell size s
}

