% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lmNormalize}
\alias{lmNormalize}
\title{Provides column by column normalisation on Y
given cell size s}
\usage{
lmNormalize(Y, s, by.class, cell.classes = NULL)
}
\arguments{
\item{Y}{Measurement by channel matrix}

\item{s}{Vector of cell sizes}

\item{by.class}{Normalize each cell class separately}

\item{cell.classes}{A vector of cell classes as each type is normalized separately. Should be the same length
as \code{nrow(Y)}, and is most likely the output of \code{cellClass(sp)} for some \code{SPData} object \code{sp}}
}
\description{
Provides column by column normalisation on Y
given cell size s
}

