\docType{methods}
\name{nCells}
\alias{nCells}
\alias{nCells,SPData-method}
\alias{nCells,SPData-methods}
\title{Number of cells in the sample}
\usage{
nCells(object)

\S4method{nCells}{SPData}(object)
}
\arguments{
  \item{object}{The SPData object to use}
}
\description{
Number of cells in the sample
}

