\docType{methods}
\name{nChannel}
\alias{nChannel}
\alias{nChannel,SPData-method}
\alias{nChannel,SPData-methods}
\title{Number of channels in the sample}
\usage{
nChannel(object)

\S4method{nChannel}{SPData}(object)
}
\arguments{
  \item{object}{The SPData object to use}
}
\description{
Returns the number of channels, which can be proteins and
protein modifications in the given sample.
}

