\docType{methods}
\name{neighbourClass,SPData,numeric-method}
\alias{neighbourClass,SPData,numeric-method}
\title{Filter out nearest neighbours by cell class. If cell i has no
nearest neighbours of class cell.class then numeric(0) is returned.}
\usage{
\S4method{neighbourClass}{SPData,numeric}(object, cell.class)
}
\description{
Filter out nearest neighbours by cell class. If cell i has
no nearest neighbours of class cell.class then numeric(0)
is returned.
}

