\name{neighbourClass}
\alias{neighbourClass}
\title{Filter out nearest neighbours by cell class. If cell i has no
nearest neighbours of class cell.class then numeric(0) is returned.}
\usage{
neighbourClass(sp, cell.class)
}
\description{
Filter out nearest neighbours by cell class. If cell i has
no nearest neighbours of class cell.class then numeric(0)
is returned.
}

