% Generated by roxygen2 (4.0.1): do not edit by hand
\name{NormalizeSP}
\alias{NormalizeSP}
\title{Normalization of an \code{SPData} object}
\usage{
NormalizeSP(sp, by.class = TRUE, norm.size = TRUE,
  norm.concentration = TRUE, norm.centrescale = TRUE,
  norm.regenerateNN = TRUE)
}
\arguments{
\item{sp}{The \code{SPData} instance to use}

\item{by.class}{Normalize each class of cells separately}

\item{norm.size}{Regress out a cell size dependency}

\item{norm.concentration}{Regress out the cell concentration dependency}

\item{norm.centrescale}{Scale each column of \code{cells(sp)} to have mean 0 and sd 1}

\item{norm.regenerateNN}{Regenerate nearest neighbour data after normalization}
}
\value{
An object of class \code{SPData} that has been appropriately normalized.
}
\description{
This method performs four steps of of normalization:
\enumerate{
\item Normalize by cell size
\item Normalize by cell concentration
\item Scale-centre the rows
\item Regenerate the nearest neighbour data from the scale centred data using the
nearest neighbour IDs
}
}
\details{
Each step of normalization can be optionally turned on or off using the appropriate flag.

Setting \code{by.class} to true means each cell class is normalized separately within a sample,
while if set to false all cells are normalized together. Note if there are many cell classes
and few cells, normalization may become nonsensical - this is not checked for.

\code{norm.concentration} constructs a measure of cell concentration for each channel as the
sum across all other channels. It then regresses each channel against the measure of concentration
and accepts the residuals as the normalized quantity.

\code{norm.regenerateNN} reconstructs the neighbour data from the newly-normalized cell data.
The neighbour information is stored separately in an \code{SPData} object so each can be
manipulated / normalized without affecting the other.
}

