% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotLm}
\alias{plotLm}
\title{Plot the linear fit of cell measurements against an independent variable (e.g. size.)}
\usage{
plotLm(Y, s, channel, cell.classes)
}
\arguments{
\item{Y}{A cell-by-channel matrix of measurements}

\item{s}{The independent (predictor) variable for plotting}

\item{cell.classes}{The classes of each cell. Should be a vector the same size}

\item{channel}{The channel to plot
as \code{nrow(Y)}.}
}
\description{
Plot the linear fit of cell measurements against an independent variable (e.g. size.)
}

