\docType{methods}
\name{rawData}
\alias{rawData}
\alias{rawData,SPData-method}
\alias{rawData,SPData-methods}
\title{Raw data matrix}
\usage{
rawData(object)

\S4method{rawData}{SPData}(object)
}
\arguments{
  \item{object}{The SPData object to use}
}
\description{
Returns the n by p matrix of raw measurements for n cells
and p channels.
}

