\name{separatenieghbours}
\alias{separatenieghbours}
\title{Separates out the nearest neighbours into only those
of a desired class, given the list X, list of nearest
neighbour ids nn.ids and the list of indices in desired
class (classList)}
\usage{
separatenieghbours(X, nnIds, classList)
}
\description{
Separates out the nearest neighbours into only those of a
desired class, given the list X, list of nearest neighbour
ids nn.ids and the list of indices in desired class
(classList)
}

