% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{SPE_report}
\alias{SPE_report}
\title{SPExp object with 5 samples used in report.}
\format{An SPExp object with 5 samples}
\usage{
data(SPE_report)
}
\description{
A dataset of 5 samples of type SPData held in an SPExp object.
}
\keyword{SPExp}

