% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{SPExp-class}
\alias{SPExp}
\alias{SPExp-class}
\title{Class SPExp}
\description{
Class \code{SPExp} defines an entire spatial proteomics experiment.
It stores a list of the samples (of type \code{SPData}) as well as
the original file names and directories.
}

