% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SPExperimentfromDir}
\alias{SPExperimentfromDir}
\title{Load an experiment from cytobank in matlab format}
\usage{
SPExperimentfromDir(directory, files = NULL)
}
\arguments{
\item{directory}{The directory containing the experiment files}

\item{files}{The filenames to load. Default is NULL, in which case all files
in the directory are used}
}
\value{
A \code{SPExp} object created from the experiment directory.
}
\description{
Load an experiment from cytobank in matlab format
}

