\name{totalProteinNormalise}
\alias{totalProteinNormalise}
\title{Uses lm normalisation to normalise by cell concentration}
\usage{
totalProteinNormalise(Y, separate = TRUE, cell.classes = NULL)
}
\arguments{
  \item{separate}{Normalise each cell class separately}

  \item{cell.classes}{The cell classes as passed by
  cellClass}
}
\description{
Uses lm normalisation to normalise by cell concentration
}

