% Generated by roxygen2 (4.0.1): do not edit by hand
\name{TotalProteinNormalize}
\alias{TotalProteinNormalize}
\title{Uses lm normalisation to normalize by cell concentration}
\usage{
TotalProteinNormalize(Y, by.class, cell.classes = NULL)
}
\arguments{
\item{Y}{A cell-by-channel matrix of measurements}

\item{by.class}{If true normalize each cell class separately}

\item{cell.classes}{The cell classes as passed by cellClass}
}
\value{
Concentration normalized cell-by-channel matrix
}
\description{
Uses lm normalisation to normalize by cell concentration
}

