\name{VIF}
\alias{VIF}
\title{Most Variance inflation factor (VIF) calculations involve supplying
a linear model, when in fact only the matrix of predictors is required.
VIF returns the variance inflation factors for predictors using only the
matrix of predictors}
\usage{
VIF(X)
}
\arguments{
  \item{X}{Predictor matrix}
}
\description{
Most Variance inflation factor (VIF) calculations involve
supplying a linear model, when in fact only the matrix of
predictors is required. VIF returns the variance inflation
factors for predictors using only the matrix of predictors
}

