% Generated by roxygen2 (4.0.1): do not edit by hand
\name{BindSPE}
\alias{BindSPE}
\title{Bind multiple sample data together}
\usage{
BindSPE(SPE, pickTumour = TRUE, useWeights = TRUE, normalise = TRUE)
}
\arguments{
\item{SPE}{The SPExp object to use}

\item{pickTumour}{Logical indicating whether or not to separate out cell types}

\item{useWeights}{Passed to neighbourMeans}

\item{normalise}{Whether to centre the columns}
}
\description{
Given an SPExp object, separate out each cell type,
calculate neighbour means and bind together, providing
more statistical power. Returns a list with cell readouts (Y),
neighbour means (X) and the number of cells from each sample (sizes)
}

