\name{neighbourChannel}
\alias{neighbourChannel}
\title{Selects only particular channels to be returned as nearest neighbours}
\usage{
neighbourChannel(NN, channel.ids)
}
\arguments{
  \item{NN}{Neighbour list (e.g. via neighbours(sp))}

  \item{channel.ids}{A channel list to select out}
}
\description{
Selects only particular channels to be returned as nearest
neighbours
}

