% Generated by roxygen2 (4.0.1): do not edit by hand
\name{neighbourChannel}
\alias{neighbourChannel}
\title{Selects only particular channels to be returned as nearest neighbours}
\usage{
neighbourChannel(NN, channel.ids)
}
\arguments{
\item{NN}{The list of matrices of neighbour data (e.g. from \code{neighbours(sp)} )}

\item{channel.ids}{A channel list to select out}
}
\description{
Selects only particular channels to be returned as nearest neighbours
}

